Move command logic into their own files
