+++ src/client/commands/Ping.js
@@ -1,0 +1,4 @@
+export default async function pingCmd (interaction) {
+  console.info('::pingCmd()')
+  await interaction.reply('pong!')
+}
