+++ src/client/commands/User.js
@@ -1,0 +1,4 @@
+export default async function userCmd (interaction) {
+  console.info('::userCmd()')
+  await interaction.reply('Not implemented yet...')
+}
